+++ dingo-common/src/main/java/io/dingodb/common/type/converter/StrParseConverter.java
@@ -55,11 +55,11 @@ public class StrParseConverter implements DataConverter {
     }
 
     @Override
-    public Double convertFloatFrom(@NonNull Object value) {
+    public Float convertFloatFrom(@NonNull Object value) {
         if (value.toString().trim().isEmpty()) {
             return null;
         }
-        return Double.parseDouble(value.toString());
+        return Float.parseFloat(value.toString());
     }
 
     @Override
